requirements.txt
pytesseract~=0.3.13
pdf2image~=1.17.0
pillow~=11.0.0
opencv-python~=4.10.0.84
Levenshtein~=0.26.0
regex~=2024.9.11
pyspellchecker~=0.8.1
fuzzywuzzy~=0.18.0
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.115.3
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.3
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe~=3.0.2
mdurl==0.1.2
numpy~=2.1.2
openpyxl==3.1.5
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
rich==13.9.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.41.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
watchfiles==0.24.0
websockets==13.1
psycopg2-binary~=2.9.10
pydantic-settings~=2.6.0
alembic~=1.13.3
pytest~=8.3.3
pydevd-pycharm~=243.21565.199
Faker~=33.1.0
gunicorn==23.0.0
uvicorn==0.34.0